// src/enums/printer.ts
export enum TextAlignment {
  LEFT = 'LEFT',
  CENTER = 'CENTER',
  RIGHT = 'RIGHT',
}

export enum TypefaceStyle {
  NORMAL = 'NORMAL',
  BOLD = 'BOLD',
  ITALIC = 'ITALIC',
  BOLD_ITALIC = 'BOLD_ITALIC',
}

export enum FontFamily {
  DEFAULT = 'DEFAULT',       
  MONOSPACE = 'MONOSPACE',
  SANS_SERIF = 'SANS_SERIF',
  SERIF = 'SERIF',
  CONDENSED = 'CONDENSED',   
}

export enum DividerStyle {
  SOLID = 'SOLID',
  DASHED = 'DASHED',
  DOTTED = 'DOTTED',
}

export enum BarcodeSymbology {
  CODE128 = 'CODE128',
  EAN13 = 'EAN13',
  CODE39 = 'CODE39',
}

export enum ImageSourceType {
  BASE64 = 'BASE64',
  PATH = 'PATH',
  URI = 'URI',
}